Version 1.0.8 finalised
Style for Registration and Login is complete. The style for every page is also complete apart for the Energy Tracker Index, Create and Delete forms.

The Index and Delete forms need to be modified for every function.

Next step Version 1.0.9

Make sure that the controllers are modified for Calculator, Energy Tracker and Schedule Installation (Schedule Consultation is already made)

The Index methods should not show the UserId instead show the Email of the user using ViewBag. Make sure that userManager is injected into the controllers

Next modify the Index so that only users and professionals see only what they need and admins can see everything

Need to seed data for the different functions (calculator, energy tracker, consultation and installation) x5 for each one.

Version 2.0.0 Should be the final version. Fully Working with no errors